Added a dockerfile

+++ libs/Mesh/tests/LineOrder2.cpp
@@ -12,10 +12,10 @@ TEST(MeshTest, LineOrder2) {
 
     LineOrder2 line({0, 1, 2}, nodes);
 
-    for (Integer ii = 0; ii < nodes->size(); ++ii) {
+    for (Integer ii = 0; ii < static_cast<Integer>(nodes->size()); ++ii) {
         EXPECT_DOUBLE_EQ(line.ShapeFn(ii, (*nodes)[static_cast<size_t>(ii)]), 1.0) << ", ii = " << ii;
 
-        for (Integer jj = 0; jj < nodes->size(); ++jj) {
+        for (Integer jj = 0; jj < static_cast<Integer>(nodes->size()); ++jj) {
             if (ii == jj) {
                 continue;
             }
@@ -25,11 +25,11 @@ TEST(MeshTest, LineOrder2) {
         }
     }
 
-    for (Integer ii = 0; ii < nodes->size(); ++ii) {
+    for (Integer ii = 0; ii < static_cast<Integer>(nodes->size()); ++ii) {
         auto sum = 0.0;
         sum += line.ShapeFn(ii, (*nodes)[static_cast<size_t>(ii)]);
 
-        for (Integer jj = 0; jj < nodes->size(); ++jj) {
+        for (Integer jj = 0; jj < static_cast<Integer>(nodes->size()); ++jj) {
             if (ii == jj) {
                 continue;
             }
@@ -52,7 +52,7 @@ TEST(MeshTest, LineOrder2ShapeFnDerivatives) {
 
     LineOrder2 line({0, 1, 2}, nodes);
 
-    for (Integer ii = 0; ii < nodes->size(); ++ii) {
+    for (Integer ii = 0; ii < static_cast<Integer>(nodes->size()); ++ii) {
         constexpr auto eps = 1.0e-8;
 
         Coord p1 = {.x = 0.0, .y = 0.0, .z = 0.0};
